Rename ecom/ui/vault/forage -> ecom/ui/vault/rosetta
All of the moving parts use the prefix rosetta instead of forage. For
example:
* RosettaPinElement
* RosettaPinSubmitter
etc.

Also, calling an implementation detail "forage" is unproductive because
all the implementation details are forage-related
